+++ test/micro-distribution/lib/Standard/Base/0.0.0-dev/src/Errors/Common.enso
@@ -25,7 +25,3 @@ type Arity_Error
 @Builtin_Type
 type Forbidden_Operation
     Error operation
-
-@Builtin_Type
-type Not_Implemented
-    Error operation
