+++ test/micro-distribution/lib/Standard/Base/0.0.0-dev/src/Runtime.enso
@@ -5,6 +5,8 @@ import project.Errors.Common.Forbidden_Operation
 import project.Function.Function
 import project.Panic.Panic
 
+from project.Runtime.Context import Input,Output
+
 @Builtin_Type
 type Context
     Input
